requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
pillow>=10.0.0
numpy>=1.26.0
scipy>=1.11.0
opencv-python-headless>=4.8.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
open-clip-torch>=2.20.0
ftfy>=6.1.0
regex>=2023.0.0
tqdm>=4.66.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
huggingface-hub>=0.20.0
safetensors>=0.4.0
timm>=0.9.0
